docs: rework repo into a structured Java Fundamentals handbook
- conspectus README per chapter: theory notes, file map, pitfalls
- root README: handbook cover with table of contents and legend
- Dutch teaching comments across all examples; IDE template junk removed
- bug/typo fixes in examples (non-compiling classes, wrong prompts, dead code)
- gap-filling examples per chapter under extra/ (IO, enums, protected/final, etc.)
- ervaringsweek: Snake game properly named and commented

+++ 04-arrays-en-enums/src/be/intecbrussel/oefening2/Oefening3.java
@@ -1,5 +1,9 @@
 package be.intecbrussel.oefening2;
 
+/**
+ * Oefening: het gemiddelde van een array berekenen met een while-lus.
+ * Valkuil: int / int geeft een afgekapt geheel getal, dus eerst casten naar double.
+ */
 public class Oefening3 {
     public static void main(String[] args) {
         int[] numbers = {20, 30, 25, 35, -16, 60, -100};
@@ -11,8 +15,9 @@ public class Oefening3 {
             i++;
         }
 
-        double average = sum / numbers.length; // 7.0 (want sum en length zijn integers)
+        // Zonder de cast zou 54 / 7 = 7 zijn (integer-deling) in plaats van 7.71...
+        double average = (double) sum / numbers.length;
 
-        System.out.println(average); // 7.0
+        System.out.println(average);
     }
 }
